Minor interface cleanup
