+++ library/src/main/kotlin/com/devbrackets/android/playlistcore/listener/MediaStatusListener.kt
@@ -3,7 +3,7 @@ package com.devbrackets.android.playlistcore.listener
 import android.support.annotation.IntRange
 import com.devbrackets.android.playlistcore.api.MediaPlayerApi
 import com.devbrackets.android.playlistcore.event.MediaProgress
-import com.devbrackets.android.playlistcore.manager.PlaylistItem
+import com.devbrackets.android.playlistcore.api.PlaylistItem
 
 /**
  * Interface definition of a callback to be invoked indicating
