+++ library/src/main/kotlin/com/devbrackets/android/playlistcore/manager/ListPlaylistManager.kt
@@ -18,6 +18,7 @@ package com.devbrackets.android.playlistcore.manager
 
 import android.app.Application
 import android.support.annotation.IntRange
+import com.devbrackets.android.playlistcore.api.PlaylistItem
 
 /**
  * An implementation of the [BasePlaylistManager] that supports Lists
